+++ src/test/java/orderbook/OrderBookTest.java
@@ -15,7 +15,7 @@ public class OrderBookTest {
     @Test
     public void shouldAddToOrderBookWithAGivenOrder() {
         OrderBook orderBook = new OrderBook();
-        Order order = new Order(Side.BID, 77.2, 1000000);
+        Order order = new Order(Side.BID, 77.2, 1000000L);
         orderBook.addOrder(order);
       /*
        * inserting to orderMap (HashMap) = O(1) usually unless hashcode collision then worst case can be O(n)
@@ -29,7 +29,7 @@ public class OrderBookTest {
     @Test
     public void shouldRemoveOrderFromBookByOrderId() {
         OrderBook orderBook = new OrderBook();
-        Order order = new Order(Side.BID, 77.2, 1000000);
+        Order order = new Order(Side.BID, 77.2, 1000000L);
         orderBook.addOrder(order);
         orderBook.removeOrderByOrderId(order.getId());
       /*
@@ -43,9 +43,9 @@ public class OrderBookTest {
     @Test
     public void shouldModifyOrderQuantityWithOrderIdAndNewQuantity() {
         OrderBook orderBook = new OrderBook();
-        Order order = new Order(Side.BID, 77.2, 1000000);
+        Order order = new Order(Side.BID, 77.2, 1000000L);
         orderBook.addOrder(order);
-        int newQuantity = 2000000;
+        Long newQuantity = 2000000L;
         orderBook.modifyOrderQuantityWithId(order.getId(), newQuantity);
       /*
        * retrieving order from orderMap (HashMap) = O(1) usually unless hashcode collision then worst case can be O(n)
@@ -58,12 +58,12 @@ public class OrderBookTest {
     @Test
     public void shouldReturnPriceOfThePositionOfLadderWithSideAndLevel() {
         OrderBook orderBook = new OrderBook();
-        Order order = new Order(Side.BID, 77.1, 1000000);
+        Order order = new Order(Side.BID, 77.1, 1000000L);
         orderBook.addOrder(order);
-        orderBook.addOrder(new Order(Side.BID, 77.0, 2000000));
-        orderBook.addOrder(new Order(Side.OFFER, 77.5, 2000000));
-        orderBook.addOrder(new Order(Side.BID, 77.2, 3000000));
-        orderBook.addOrder(new Order(Side.BID, 76.9, 3000000));
+        orderBook.addOrder(new Order(Side.BID, 77.0, 2000000L));
+        orderBook.addOrder(new Order(Side.OFFER, 77.5, 2000000L));
+        orderBook.addOrder(new Order(Side.BID, 77.2, 3000000L));
+        orderBook.addOrder(new Order(Side.BID, 76.9, 3000000L));
         assertThat(orderBook.getPriceWithSideAndLevel(Side.BID, 1), equalTo(order.getPrice()));
       /*
        * getting entrySet from ladder
@@ -75,13 +75,13 @@ public class OrderBookTest {
     @Test
     public void shouldReturnTotalQuantityWithSideAndLevel() {
         OrderBook orderBook = new OrderBook();
-        orderBook.addOrder(new Order(Side.BID, 77.2, 3000000));
-        Order order1 = new Order(Side.BID, 77.1, 1000000);
-        Order order2 = new Order(Side.BID, 77.1, 3000000);
+        orderBook.addOrder(new Order(Side.BID, 77.2, 3000000L));
+        Order order1 = new Order(Side.BID, 77.1, 1000000L);
+        Order order2 = new Order(Side.BID, 77.1, 3000000L);
         orderBook.addOrder(order1);
         orderBook.addOrder(order2);
-        orderBook.addOrder(new Order(Side.BID, 77.0, 2000000));
-        orderBook.addOrder(new Order(Side.OFFER, 77.5, 2000000));
+        orderBook.addOrder(new Order(Side.BID, 77.0, 2000000L));
+        orderBook.addOrder(new Order(Side.OFFER, 77.5, 2000000L));
         assertThat(orderBook.getTotalQuantityWithSideAndLevel(Side.BID, 1),
                 equalTo(order1.getQuantity() + order2.getQuantity()));
        /*
